+++ gormx/migrate/testdata/migrations/00002_add_widget_color.sql
@@ -1,0 +1,5 @@
+-- +goose Up
+ALTER TABLE widgets ADD COLUMN color TEXT;
+
+-- +goose Down
+ALTER TABLE widgets DROP COLUMN color;